Change value/price from int to long as defined in the interfaces.

+++ basketserver/src/main/java/de/upb/crypto/incentive/basketserver/model/requests/PayBasketRequest.java
@@ -14,5 +14,5 @@ import java.util.UUID;
 @Data
 public class PayBasketRequest {
     UUID basketId;
-    int value;
+    long value;
 }
